backend/ai-services/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
python-docx>=1.1.0
markdown>=3.5.0
pymilvus>=2.3.0
elasticsearch>=8.12.0
httpx>=0.26.0
tenacity>=8.2.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
